feature(Events): add ability to specify aggregate function
