+++ src/fold/events/labeling/fixed.py
@@ -1,10 +1,11 @@
 from __future__ import annotations
 
-from typing import List, Optional
+from typing import Callable, List, Optional, Union
 
 import pandas as pd
 
-from ...utils.forward import create_forward_rolling_sum
+from ...base import PredefinedFunction
+from ...utils.forward import create_forward_rolling
 from ..base import EventDataFrame, Labeler, LabelingStrategy, WeighingStrategy
 from ..weights import NoWeighing, WeightBySumWithLookahead
 
@@ -18,6 +19,9 @@ class FixedForwardHorizon(Labeler):
         labeling_strategy: LabelingStrategy,
         weighing_strategy: Optional[WeighingStrategy],
         weighing_strategy_test: WeighingStrategy = WeightBySumWithLookahead(),
+        aggregate_function: Union[
+            PredefinedFunction, Callable
+        ] = PredefinedFunction.sum,
     ):
         self.time_horizon = time_horizon
         self.labeling_strategy = labeling_strategy
@@ -25,17 +29,20 @@ class FixedForwardHorizon(Labeler):
             weighing_strategy if weighing_strategy else NoWeighing()
         )
         self.weighing_strategy_test = weighing_strategy_test
+        self.aggregate_function = aggregate_function
 
     def label_events(
         self, event_start_times: pd.DatetimeIndex, y: pd.Series
     ) -> EventDataFrame:
-        forward_rolling_sum = create_forward_rolling_sum(y, self.time_horizon)
+        forward_rolling_aggregated = create_forward_rolling(
+            self.aggregate_function, y, self.time_horizon
+        )
         cutoff_point = y.index[-self.time_horizon]
         event_start_times = event_start_times[event_start_times < cutoff_point]
-        event_candidates = forward_rolling_sum[event_start_times]
+        event_candidates = forward_rolling_aggregated[event_start_times]
 
         labels = self.labeling_strategy.label(event_candidates)
-        raw_returns = forward_rolling_sum[event_start_times]
+        raw_returns = forward_rolling_aggregated[event_start_times]
         sample_weights = self.weighing_strategy.calculate(raw_returns)
         test_sample_weights = self.weighing_strategy_test.calculate(raw_returns)
